django_microservice/pyproject.toml
[project]
name = "itmo"
version = "0.1.0"
description = ""
authors = [
    {name = "cerg",email = "cerg0909@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "locust (>=2.43.0,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "adrf (>=0.1.12,<0.2.0)",
    "orjson (>=3.10,<4.0.0)",
    "psutil (>=5.9,<7.0.0)",
    "psycopg-pool (>=3.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
